change type of margin, create function to set margins for view

+++ projectsup/src/main/java/com/ludaxord/projectsup/library/button/button/floatingbutton/SupFloatingButton.kt
@@ -28,7 +28,7 @@ class SupFloatingButton : AbstractSupFloatingButton {
         setCornersFromFloat(corners)
     }
 
-    override fun setStickToBottom(withMargin: Float, view: View?) {
+    override fun setStickToBottom(withMargin: Int, view: View?) {
         setStickToBottomOfView(withMargin)
     }
